requirements.txt
aiogram==3.26.0
APScheduler==3.11.1
python-dotenv>=1.0.1
